Create a Constant

Create either a "scalar" (hlo_scalar) or tensor (hlo_tensor) constant. Strictly speaking, stableHLO "scalars" are simply tensors with 0 dimensions. To create an emtpy constant (at least one dimension is 0), use hlo_empty.

Usage

hlo_scalar(value, ..., func = .current_func())

# S3 method for class 'logical'
hlo_scalar(value, ..., func = .current_func())

# S3 method for class 'double'
hlo_scalar(value, ..., dtype = NULL, func = .current_func())

# S3 method for class 'integer'
hlo_scalar(value, ..., dtype = NULL, func = .current_func())

# S3 method for class 'PJRTBuffer'
hlo_scalar(value, ..., func = .current_func())

hlo_tensor(value, ..., func = .current_func())

# S3 method for class 'array'
hlo_tensor(value, ..., dtype = NULL, func = .current_func())

# S3 method for class 'integer'
hlo_tensor(value, ..., dtype = NULL, shape = NULL, func = .current_func())

# S3 method for class 'logical'
hlo_tensor(value, ..., dtype = NULL, shape = NULL, func = .current_func())

# S3 method for class 'double'
hlo_tensor(value, ..., dtype = NULL, shape = NULL, func = .current_func())

# S3 method for class 'PJRTBuffer'
hlo_tensor(value, ..., func = .current_func())

hlo_empty(dtype, shape, func = .current_func())

Arguments

value

(any)
Value from which to create a constant.

...

(any)
Additional arguments including:

dtype (character(1)): String for element type. Can be one of f64, f32, u8, u16, u32, u64, i8, i16, i32, i64, pred.

shape (integer()): Shape of the tensor (for hlo_tensor only). If not specified, the shape is inferred from the data.

func

(Func)
The function to add the constant to. Per default, uses the last function created with hlo_func or local_func.

dtype

(character(1))
String for element type. Can be one of f64, f32, u8, u16, u32, u64, i8, i16, i32, i64, pred.

shape

(integer())
Shape of the tensor (for hlo_tensor only). If not specified, the shape is inferred from the data.

Examples

hlo_scalar(1L, dtype = "i32", func = Func())
#> Variable %0 in:
#> func.func @ () ->  {
#> %0 = "stablehlo.constant" () {
#> value = dense<1> : tensor<i32>
#> }: () -> (tensor<i32>)
#> }
hlo_scalar(1, dtype = "f32", func = Func())
#> Variable %0 in:
#> func.func @ () ->  {
#> %0 = "stablehlo.constant" () {
#> value = dense<1.00000000e+00> : tensor<f32>
#> }: () -> (tensor<f32>)
#> }
hlo_scalar(TRUE, func = Func())
#> Variable %0 in:
#> func.func @ () ->  {
#> %0 = "stablehlo.constant" () {
#> value = dense<true> : tensor<i1>
#> }: () -> (tensor<i1>)
#> }
hlo_tensor(array(c(1, 2, 3, 4), dim = c(1, 4)), dtype = "f32", func = Func())
#> Variable %0 in:
#> func.func @ () ->  {
#> %0 = "stablehlo.constant" () {
#> value = dense<[[1.00000000e+00, 2.00000000e+00, 3.00000000e+00, 4.00000000e+00]]> : tensor<1x4xf32>
#> }: () -> (tensor<1x4xf32>)
#> }
hlo_empty(dtype = "f32", shape = c(0, 3), func = Func())
#> Variable %0 in:
#> func.func @ () ->  {
#> %0 = "stablehlo.constant" () {
#> value = dense<[]> : tensor<0x3xf32>
#> }: () -> (tensor<0x3xf32>)
#> }